Revert "Improve Yocto integration"
This reverts commit 877ffdef84ac2cd59312b28d4e0258b5d07111e1.

The use of Yocto images is not suitable for all target builds, such as
vm_minimal, vm_multi etc. Where Yocto images are required, the project
should point directly to the location of the images (i.e. make
assumption for the location), and have means for overriding the defaults.

See https://github.com/tiiuae/tii-sel4-vm/pull/23 for example.